relationships: fix trying to preload for removed entity with invalid relationship [closes #434]
This approach moves nullability check from IProperty.getRawValue() to
Entity.getRawValues() method which is used when storing an entity.

Other approach may be exposing getPrimaryValue() method with same
functionality but the null check. This alternative apporach seems to be
more intrusive to current API of properties & relationships. But we may
reconsider later.