dbal builder: rewrite WHERE condition for has-many rel into JOIN clause and aggretaion in HAVING
This commit changes how we handle has-many relationship query construciton.
Until now, every has-many relationship was simply left-joined and filtered
via where. The duplication was solved by group-by statement.

From now on, the joined table is filtered already in the JOIN clause and the
where-part is simply aggretation count of the joined table > 0.

This allow moving the where clause to having clause, where it may be
combined with other aggregation filtering condition.